+++ src/threedidepth/commands.py
@@ -13,12 +13,12 @@ from threedidepth.calculate import MODE_CONSTANT
 def threedidepth(*args):
     parser = argparse.ArgumentParser()
     parser.add_argument(
-        "gridadmin_path", metavar="gridadmin", help="path to gridadmin file"
+        "gridadmin_path", metavar="gridadmin", help="path to gridadmin.h5 file"
     )
     parser.add_argument(
         "results_3di_path",
         metavar="results_3di",
-        help="path to simulation results file",
+        help="path to (aggregate_)results_3di.nc file",
     )
     parser.add_argument(
         "dem_path", metavar="dem", help="path to bathymetry file"
@@ -33,7 +33,6 @@ def threedidepth(*args):
         "--steps",
         nargs="+",
         type=int,
-        default=[-1, ],
         dest="calculation_steps",
         help="simulation result step(s)",
     )